chore: address review comments on the rebuilt #134
- interface: rename the unused DEFAULT_POOL_TREE_PUBKEY to DEFAULT_TREE_ADDRESS
  and move the from_account_bytes unit test into tests/ (add dev-deps).
- client: document that to_merge_proof packs the same P256 five-tuple as
  to_transact_proof but as the raw [u8;192] the merge_transact ix requires.
- spec: correct ViewingKey::from_seed to the implemented flat HKDF derivation
  and add the ed25519 SigningKey constructors; mark SLIP-0010-P256 as the
  planned, not-yet-implemented P-256 rail.
