+++ program-libs/interface/src/lib.rs
@@ -33,7 +33,7 @@ pub const PROGRAM_ID_PUBKEY: solana_pubkey::Pubkey =
 pub const DEFAULT_TREE_ADDRESS: &str = "treeYbr45LjxovKvtD46uEphM64kwoFFPYhVNw1A8x8";
 
 /// [`DEFAULT_POOL_TREE_ADDRESS`] as a `Pubkey`.
-pub const DEFAULT_POOL_TREE_PUBKEY: solana_pubkey::Pubkey =
+pub const DEFAULT_TREE_ADDRESS: solana_pubkey::Pubkey =
     solana_pubkey::Pubkey::from_str_const(DEFAULT_POOL_TREE_ADDRESS);
 
 /// Seed for the native SOL interface account used by public SOL settlement.
